requirements.txt
streamlit>=1.24.0
PyPDF2>=3.0.0
sentence-transformers>=2.2.5
transformers>=4.35.0
torch>=2.0.1
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
openai>=1.0.0
huggingface_hub>=0.20.0
